Use `<Markprompt />` from `@markprompt/react` in `@markprompt/docusaurus-theme-search` and update examples (#47)
* Use Markprompt in @markprompt/docusaurus-theme-search

* Update Next.js and Docusaurus examples to use the headful Markprompt component

* Update example dependencies

* Build @markprompt/docusaurus-theme-search in CI

* Add paths to @markprompt/docusaurus-theme-search tsconfig.json

* Cancel previous runs of the CI workflow in the same branch

* Add changesets

* Add MARKPROMPT_PROJECT_KEY environment variable to CI build step

* Don't render <Markprompt /> on the server in with-docusaurus-swizzled

* Add build/ to .prettierignore

* Add build/ to .eslintrc.js ignore
