+++ .changeset/olive-days-juggle.md
@@ -1,0 +1,5 @@
+---
+'@markprompt/docusaurus-theme-search': minor
+---
+
+Use the prebuilt Markprompt component from `@markprompt/react` in `@markprompt/docusaurus-theme-search`
