+++ .changeset/unlucky-berries-marry.md
@@ -1,0 +1,5 @@
+---
+'@markprompt/css': patch
+---
+
+Make sure dark/light themes work for Docusaurus
