+++ .github/workflows/ci.yml
@@ -5,6 +5,10 @@ on:
   push:
     branches: [main]
 
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: ${{ github.ref != 'refs/heads/main' }}
+
 jobs:
   lint:
     runs-on: ubuntu-latest
@@ -15,8 +19,10 @@ jobs:
           node-version: 18
           cache: npm
       - run: npm ci
-      - run: npm run build -w @markprompt/core -w @markprompt/react -w @markprompt/web
+      - run: npm run build -w @markprompt/core -w @markprompt/docusaurus-theme-search -w @markprompt/react -w @markprompt/web
       - run: npm run build -w with-css-modules -w with-docusaurus -w with-docusaurus-swizzled -w with-next -w with-markprompt-web
+        env:
+          MARKPROMPT_PROJECT_KEY: ${{ secrets.MARKPROMPT_PROJECT_KEY }}
       - run: npm run lint
 
   test:
