+++ examples/with-docusaurus-swizzled/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@markprompt/with-docusaurus-swizzled",
+  "name": "with-docusaurus-swizzled",
   "version": "0.0.2",
   "private": true,
   "scripts": {
@@ -32,7 +32,8 @@
     "@docusaurus/preset-classic": "^2.4.0",
     "@docusaurus/theme-search-algolia": "^2.4.1",
     "@markprompt/css": "*",
-    "@markprompt/web": "*",
+    "@markprompt/docusaurus-theme-search": "^0.3.0",
+    "@markprompt/react": "*",
     "@mdx-js/react": "^1.6.22",
     "prism-react-renderer": "^1.3.5",
     "react": "^17.0.2",
