+++ examples/with-docusaurus/docusaurus.config.js
@@ -42,6 +42,17 @@ const config = {
         // the project key in the Markprompt dashboard, under
         // project settings.
         projectKey: process.env.MARKPROMPT_PROJECT_KEY,
+        trigger: { floating: true },
+        references: {
+          transformReferenceId: (referenceId) => {
+            // Sample code that transforms a reference path to a link.
+            // Remove file extension
+            const href = referenceId.replace(/\.[^.]+$/, '');
+            // Use last part of path for label
+            const text = href.split('/').slice(-1)[0];
+            return { text, href };
+          },
+        },
       },
       navbar: {
         title: 'Markprompt',
