+++ examples/with-docusaurus/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@markprompt/with-docusaurus",
+  "name": "with-docusaurus",
   "version": "0.0.2",
   "private": true,
   "scripts": {
