+++ examples/with-markprompt-web/src/main.ts
@@ -7,7 +7,7 @@ const el = document.querySelector('#markprompt');
 
 if (el && el instanceof HTMLElement) {
   markprompt(import.meta.env.VITE_PROJECT_API_KEY, el, {
-    trigger: { floating: true },
+    trigger: { floating: false },
     search: { enable: true },
   });
 }
