+++ examples/with-next/.env
@@ -1,1 +1,1 @@
-MARKPROMPT_PROJECT_KEY='Your Markprompt project key'
+NEXT_PUBLIC_MARKPROMPT_PROJECT_KEY='Your Markprompt project key'
