+++ examples/with-next/pages/_app.tsx
@@ -1,8 +1,8 @@
 import '@markprompt/css';
+import './global.css';
+
 import { AppProps } from 'next/app';
 import React, { ReactElement } from 'react';
-
-import './global.css';
 
 export default function App({ Component, pageProps }: AppProps): ReactElement {
   return <Component {...pageProps} />;
