+++ packages/css/markprompt.css
@@ -21,7 +21,8 @@
 }
 
 @media (prefers-color-scheme: dark) {
-  :root {
+  /* Support Docusaurus dark theme data attribute */
+  :not([data-theme='light']):root {
     --markprompt-background: #050505;
     --markprompt-foreground: #d4d4d4;
     --markprompt-muted: #171717;
@@ -38,6 +39,25 @@
     --markprompt-overlay: #00000040;
     --markprompt-ring: #ffffff;
   }
+}
+
+/* Support Docusaurus dark theme data attribute */
+[data-theme='dark']:root {
+  --markprompt-background: #050505;
+  --markprompt-foreground: #d4d4d4;
+  --markprompt-muted: #171717;
+  --markprompt-mutedForeground: #737373;
+  --markprompt-border: #262626;
+  --markprompt-input: #ffffff;
+  --markprompt-primary: #6366f1;
+  --markprompt-primaryForeground: #ffffff;
+  --markprompt-primaryMuted: #8285f4;
+  --markprompt-secondary: #0e0e0e;
+  --markprompt-secondaryForeground: #ffffff;
+  --markprompt-primaryHighlight: #ec4899;
+  --markprompt-secondaryHighlight: #a855f7;
+  --markprompt-overlay: #00000040;
+  --markprompt-ring: #ffffff;
 }
 
 :where([class^='Markprompt']) {
